Add files for top 25 viewed articles

+++ src/main/java/reducers/CovidReduce.java
@@ -10,7 +10,7 @@ public class CovidReduce extends Reducer<Text, LongWritable, Text, LongWritable>
     public void reduce(Text key, Iterable<LongWritable> values, Context context)
             throws IOException, InterruptedException {
 
-        long views = (long) 0;
+        long views = 0;
         for (LongWritable t : values) {
             views += t.get();
         }
